Rename cassandra python package
It conflicted with cassandra-driver.
