+++ AppDB/cassandra_env/shut_down_cassandra.py
