+++ AppDB/cassandra_env/templates/cassandra-env.sh
