+++ AppDB/cassandra_env/templates/cassandra.in.sh
