+++ AppDB/cassandra_env/templates/cassandra.yaml
