+++ AppDB/cassandra_env/templates/log4j-server.properties
