+++ AppDB/cassandra_env/thrift_cass/Cassandra-remote
