+++ AppDB/cassandra_env/thrift_cass/constants.py
