+++ AppDB/cassandra_env/thrift_cass/remote
