+++ AppDB/cassandra_env/thrift_cass/ttypes.py
