+++ AppDB/test/functional/test_cassandra.py
@@ -2,7 +2,7 @@ import os
 import sys
 import unittest
 
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../cassandra"))
+sys.path.append(os.path.join(os.path.dirname(__file__), "../../cassandra_env"))
 import cassandra_interface
 
 # Prereq: Cassandra must be running first
