+++ AppDB/test/unit/test_cassandra_interface.py
@@ -8,7 +8,7 @@ import unittest
 
 from flexmock import flexmock
 
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../cassandra"))  
+sys.path.append(os.path.join(os.path.dirname(__file__), "../../cassandra_env"))
 import cassandra_interface
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "../../../lib/"))
